Fix the remaining warnings; the build is now clean
Zero warnings, zero errors, down from 46.

Counts and indices are NSUInteger/NSInteger rather than unsigned/int. The loop
counters became NSInteger rather than NSUInteger deliberately: one of them, in
AppsForItem, counts down and tests i >= 0, which an unsigned type turns into an
infinite loop. Widening -childCount and -childAtIndex: rippled out to their
callers, which is why this touches more files than the warning list suggested.

Two of these were more than cosmetic. The outline view's data source implemented
-outlineView:numberOfChildrenOfItem: returning int and -outlineView:child:ofItem:
taking int, where AppKit declares both NSInteger. On 64-bit that means the return
value is written 32 bits wide and read 64, which happens to work only while the
counts stay small. The table view delegates had the same problem with their row
arguments.

The rest were project-level:

  - The Run Script phase stripped header directories out of the embedded
    frameworks. There are no embedded frameworks any more, so it did nothing
    except run on every build and warn about it. Removed.
  - ALWAYS_SEARCH_USER_PATHS set to NO, which is what the traditional-headermap
    warning asks for.
  - Timing.c renamed to Timing.m. It was the only C file in the target, so
    -fobjc-arc-exceptions did not apply to it and clang warned that the flag was
    unused. Objective-C is a superset of C, so the file is unchanged.
  - ZoomInfo called its delegate through -performSelector:withObject:, which ARC
    warns about because it cannot know the selector's memory semantics. It now
    calls the IMP directly, which states them.

Verified: the panes, treemap, outline and hit testing all still behave, and the
app scans and displays a folder without crashing.

+++ AppsForItem.m
@@ -144,7 +144,8 @@
             result = [NSMutableArray arrayWithArray:outURLs];
             
             // filter out .exe files
-            int i, cnt = [result count];
+            //signed: this loop counts down and tests i >= 0
+            NSInteger i, cnt = [result count];
             NSURL *url;
             
             for (i=(cnt-1);i>=0;i--)
