+++ FilesOutlineViewController.m
@@ -94,7 +94,7 @@
 
 #pragma mark --------NSOutlineView datasource-----------------
 
-- (id) outlineView: (NSOutlineView *) outlineView child: (int) index ofItem: (id) item
+- (id) outlineView: (NSOutlineView *) outlineView child: (NSInteger) index ofItem: (id) item
 {
 	FSItem *fsItem = (item == nil) ? [self rootItem] : item;
 
@@ -106,7 +106,7 @@
     return [[self document] itemIsNode: item];
 }
 
-- (int) outlineView: (NSOutlineView *) outlineView numberOfChildrenOfItem: (id) item
+- (NSInteger) outlineView: (NSOutlineView *) outlineView numberOfChildrenOfItem: (id) item
 {
 	FSItem *fsItem = (item == nil) ? [self rootItem] : item;
